// src/interfaces/station.ts
interface Station {
    brand: string;
    diesel: number;
    dist: number;
    e5: number;
    e10: number;
    houseNumber: string;
    id: string;
    isOpen: boolean;
    lat: number;
    lng: number;
    name: string;
    place: string;
    postCode: number;
    street: string;
}

export default Station;